Refactor class app, add FileReader class
